# rula/docs/swagger.yaml
definitions:
  models.CreateUserReq:
    properties:
      address:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.DeleteUserReq:
    properties:
      username:
        type: string
    type: object
  models.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
  models.UpdateUserReq:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      operationId: general
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
  /login/:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login
      parameters:
      - description: LoginReq
        in: body
        name: models.LoginReq
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login
  /user/:
    delete:
      consumes:
      - application/json
      description: DeleteUser
      operationId: Delete User
      parameters:
      - description: DeleteUserReq
        in: body
        name: models.DeleteUserReq
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: DeleteUser
    post:
      consumes:
      - application/json
      description: Create new User
      operationId: create-user
      parameters:
      - description: CreateUserReq
        in: body
        name: models.CreateUserReq
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create new User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: Update User
      parameters:
      - description: UpdateUserReq
        in: body
        name: models.UpdateUserReq
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update User
swagger: "2.0"